+++ prebuild/emptyDir.js
@@ -1,17 +1,17 @@
 const fs = require('fs');
 
 module.exports = {
-    emptyFolder: path => {
+    emptyFolder: (path, deleteFolder = true) => {
         if (fs.existsSync(path)) {
             fs.readdirSync(path).forEach(file => {
                 let curPath = path + '/' + file;
                 if (fs.lstatSync(curPath).isDirectory()) {
-                    module.exports.emptyFolder(curPath);
+                    module.exports.emptyFolder(curPath, true);
                 } else {
                     fs.unlinkSync(curPath);
                 }
             });
-            fs.rmdirSync(path);
+            if (deleteFolder) fs.rmdirSync(path);
         }
     },
 };
